useRating hook
